+++ docs/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ## [Unreleased]
 
+## [0.4.0] - 2024-05-17
+
 ### Added
 
 -   Allowed operations between timings
@@ -52,7 +54,8 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
-[Unreleased]: https://github.com/spapanik/pyutilkit/compare/v0.3.0...main
+[Unreleased]: https://github.com/spapanik/pyutilkit/compare/v0.4.0...main
+[0.4.0]: https://github.com/spapanik/pyutilkit/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/spapanik/pyutilkit/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/spapanik/pyutilkit/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/spapanik/pyutilkit/releases/tag/v0.1.0